# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Church Youth Group Coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scheduled-activities:
    get:
      operationId: listScheduledActivities
      tags: [scheduled-activities]
      summary: List scheduled weekly activities for current group
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Scheduled activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledActivity"

    post:
      operationId: createScheduledActivity
      tags: [scheduled-activities]
      summary: Schedule a weekly activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledActivityInput"
      responses:
        "201":
          description: Scheduled activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledActivity"
  /scheduled-activities/{id}:
    patch:
      operationId: updateScheduledActivity
      tags: [scheduled-activities]
      summary: Update a scheduled activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledActivity"

    delete:
      operationId: deleteScheduledActivity
      tags: [scheduled-activities]
      summary: Delete a scheduled activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SUNDAY LESSONS
...
  /sacrament-rotations:
    get:
      operationId: listSacramentRotations
      tags: [sacrament]
      summary: List sacrament blessing rotations for current group
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Rotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SacramentRotation"

    post:
      operationId: createSacramentRotation
      tags: [sacrament]
      summary: Create a sacrament rotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SacramentRotationInput"
      responses:
        "201":
          description: Created rotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SacramentRotation"
...
  /sacrament-rotations/{id}:
    patch:
      operationId: updateSacramentRotation
      tags: [sacrament]
      summary: Update a sacrament rotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SacramentRotationUpdate"
      responses:
        "200":
          description: Updated rotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SacramentRotation"

    delete:
      operationId: deleteSacramentRotation
      tags: [sacrament]
      summary: Delete a sacrament rotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SUBSTITUTION REQUESTS
  /substitution-requests:
    get:
      operationId: listSubstitutionRequests
      tags: [substitutions]
      summary: List substitution requests for current group
      responses:
        "200":
          description: Substitution requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubstitutionRequest"

    post:
      operationId: createSubstitutionRequest
      tags: [substitutions]
      summary: Request a substitution for a sacrament rotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubstitutionRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubstitutionRequest"
  /substitution-requests/{id}:
    patch:
      operationId: updateSubstitutionRequest
      tags: [substitutions]
      summary: Approve, deny, or resolve a substitution request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubstitutionRequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubstitutionRequest"

  # DASHBOARD
...
components:
  schemas:
...
    ScheduledActivity:
      type: object
      required: [id, date, activityId, activityTitle, groupId]
      properties:
        id:
          type: integer
        date:
          type: string
        activityId:
          type: integer
        activityTitle:
          type: string
        activityType:
          type: string
        personInChargeId:
          type: ["integer", "null"]
        personInChargeName:
          type: ["string", "null"]
        treatsAssigneeId:
          type: ["integer", "null"]
        treatsAssigneeName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        groupId:
          type: integer
        createdAt:
          type: string
    ScheduledActivityInput:
      type: object
      required: [date, activityId]
      properties:
        date:
          type: string
        activityId:
          type: integer
        personInChargeId:
          type: ["integer", "null"]
        treatsAssigneeId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ScheduledActivityUpdate:
      type: object
      properties:
        date:
          type: string
        activityId:
          type: integer
        personInChargeId:
          type: ["integer", "null"]
        treatsAssigneeId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # SUNDAY LESSONS
...
    SacramentRotation:
      type: object
      required: [id, date, groupId, members]
      properties:
        id:
          type: integer
        date:
          type: string
        groupId:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/RotationMember"
        createdAt:
          type: string
    RotationMember:
      type: object
      required: [userId, firstName, lastName]
      properties:
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: ["string", "null"]
    SacramentRotationInput:
      type: object
      required: [date, memberIds]
      properties:
        date:
          type: string
        memberIds:
          type: array
          items:
            type: integer
          minItems: 1
          maxItems: 3
    SacramentRotationUpdate:
      type: object
      properties:
        date:
          type: string
        memberIds:
          type: array
          items:
            type: integer
...
    SubstitutionRequest:
      type: object
      required: [id, rotationId, requesterId, status, groupId]
      properties:
        id:
          type: integer
        rotationId:
          type: integer
        rotationDate:
          type: string
        requesterId:
          type: integer
        requesterName:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, denied, resolved]
        groupId:
          type: integer
        createdAt:
          type: string
    SubstitutionRequestInput:
      type: object
      required: [rotationId]
      properties:
        rotationId:
          type: integer
        reason:
          type: ["string", "null"]
    SubstitutionRequestUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, denied, resolved]

    # DASHBOARD
...