# lib/api-zod/src/generated/types/sacramentRotation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church Youth Group Coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { RotationMember } from './rotationMember';

export interface SacramentRotation {
  id: number;
  date: string;
  groupId: number;
  members: RotationMember[];
  createdAt?: string;
}